+++ connector-packager/LICENSE
