+++ connector-packager/connector_packager/connector_file.py
@@ -3,11 +3,11 @@ import os.path
 
 class ConnectorFile:
 
-    def __init__(self, file_name, file_type):
+    def __init__(self, file_name: str, file_type: str):
         self.file_name = file_name
         self.file_type = file_type
 
-    def extension(self):
+    def extension(self) -> str:
         ext = os.path.splitext(self.file_name)[-1]
         # Remove the leading period.
         return ext if not ext else ext[1:]
