+++ connector-packager/connector_packager/helper.py
