+++ connector-packager/connector_packager/xml_parser.py
