+++ connector-packager/connector_packager/xsd_validator.py
