+++ connector-packager/tests/test_package.py
@@ -4,7 +4,7 @@ import logging
 import argparse
 from pathlib import Path
 
-from package.package import main, PACKAGED_EXTENSION
+from connector_packager.package import main, PACKAGED_EXTENSION
 
 logger = logging.getLogger(__name__)
 
@@ -23,7 +23,7 @@ class TestPackage(unittest.TestCase):
             logging.debug("Removing old test file " + str(Path))
             path_to_test_file.unlink()
 
-        os.system("python package.package " + str(files_directory) +
+        os.system("python connector_packager.package " + str(files_directory) +
                   " --package-only " + " --dest " + str(expected_dest_directory))
 
         self.assertTrue(path_to_test_file.exists(),
