+++ connector-packager/tests/test_resources/big_manifest/manifest.xml
