+++ connector-packager/tests/test_resources/broken_xml/connectionResolver.tdr
