+++ connector-packager/tests/test_resources/broken_xml/manifest.xml
