+++ connector-packager/tests/test_resources/valid_connector/connection-dialog.tcd
