+++ connector-packager/tests/test_resources/valid_connector/connectionBuilder.js
