+++ connector-packager/tests/test_resources/valid_connector/dialect.tdd
