+++ connector-packager/tests/test_resources/valid_connector/manifest.xml
