+++ connector-packager/tests/test_resources/valid_connector/resources-en_US.xml
