+++ connector-packager/tests/test_resources/wrong_class/connection-dialog.tcd
