+++ connector-packager/tests/test_resources/wrong_class/manifest.xml
